+++ src/board.py
@@ -7,7 +7,7 @@ Note: Board contains all game functions that don't relate to AI
 """
 
 __all__ = ['Board', 'PiecesBar']
-__version__ = '0.2'
+__version__ = '0.3'
 __author__ = 'Eric G.D'
 
 from typing import List, Set
@@ -28,24 +28,26 @@ class Board(GridLayout):
     Class Board(GridLayout):
     ------------------------
 
-    The main game board, also returned by QuartoApp.build
+    Represents the main game board where pieces are inserted.
+    In addition, it contains most of the main program's code
     """
 
     LENGTH = 4
-    DIFFICULTY = {'baby': 0, 'easy': 4, 'medium': 8,
-                  'hard': 12, 'impossible': LENGTH ** 2}
+    DIFFICULTY = {'baby': 0, 'easy': 2, 'medium': 4,
+                  'hard': 6, 'impossible': LENGTH ** 2}
     popup = None
 
     def __init__(self, **kwargs) -> None:
         super().__init__()
         self.rows = self.cols = Board.LENGTH
-        self.first_player = self.current_player = kwargs.get('first_player', Player.HUMAN)
-        self.game_mode = kwargs.get('game_mode', GameMode.SINGLE_PLAYER)
+        self.first_player = self.current_player = kwargs.get('first_player', Player.human)
+        self.game_mode = kwargs.get('game_mode', GameMode.single_player)
         self.depth = Board.DIFFICULTY[kwargs.get('difficulty', 'hard')]
         self.cell_list = [[Cell() for _ in range(self.cols)] for _ in range(self.rows)]
-        self.pieces_bar = None
-        Logger.debug("Setup: Initialised new Board object")
-        self.disabled = True    # Don't allow game to run until self.pieces_bar is not None
+        self.pieces_bar = PiecesBar(self)
+        self.initialise_buttons()
+        self.first_move()
+        Logger.debug("Setup: Initialised new {game_mode} Board".format(game_mode=self.game_mode))
 
     def initialise_buttons(self, reset: bool = False) -> None:
         """
@@ -73,21 +75,21 @@ class Board(GridLayout):
         MULTI_PLAYER: Gets a piece from the first player and the second player's move
         :return:    None
         """
-        if self.game_mode == GameMode.SINGLE_PLAYER:
-            if self.first_player == Player.COMPUTER:
-                self.computer_move()
-            else:
-                pick_piece(self.cell_list, self.pieces_bar.pieces_set, self.depth)
+        if self.game_mode == GameMode.single_player:
+            if self.first_player == Player.human:
+                piece = pick_starting_piece(self.pieces_bar.pieces_set)
+                self.pieces_bar.confirm(piece)
 
     def computer_move(self) -> None:
         """
         Finds the best move for the computer and plays it
         :return:    None
         """
-        option = minimax(self, self.pieces_bar.confirmed, self.pieces_bar.pieces_set, self.depth)
+        option = minimax(self, self.pieces_bar.confirmed.piece, self.pieces_bar.pieces_set, self.depth)
         i, j = option.index
-        self.insert(self.cell_list[i][j], option.piece)
-        Logger.debug('Application: Computer placed {} at {}'.format(self.pieces_bar, option.index))
+        self.insert_confirmed(self.cell_list[i][j])
+        self.pieces_bar.confirm(option.piece)
+        Logger.info('Application: Computer placed {} at {}'.format(self.pieces_bar, option.index))
         self.change_player()
 
     def on_click(self, touch) -> None:
@@ -117,9 +119,13 @@ class Board(GridLayout):
         winning_move = has_won(converted)
         game_over = is_full or winning_move
         if game_over:
-            title = '{} wins!'.format(self.current_player) if winning_move else "It's a tie!"
-            self.end_message(title)
+            message = '{} wins!'.format(self.get_current_player_str()) if winning_move else "It's a tie!"
+            self.end_message(message)
         return game_over
+
+    def get_current_player_str(self) -> str:
+        return self.current_player.name.title() if self.game_mode == GameMode.single_player else \
+            f'Player {self.current_player.value}'
 
     def insert_confirmed(self, cell: Cell) -> bool:
         """
@@ -139,7 +145,8 @@ class Board(GridLayout):
         :return:        None
         """
         self.disabled = True
-        self.popup = EndMessage()
+        if self.popup is None:
+            self.popup = EndMessage()
         self.popup.ids.message.text = message
         self.popup.open()
 
@@ -150,7 +157,6 @@ class Board(GridLayout):
         """
         if self.popup is not None:
             self.popup.dismiss()
-            self.popup = None
         self.disabled = False
         self.initialise_buttons(reset=True)
         self.first_player = self.current_player = (~self.first_player)
@@ -193,26 +199,24 @@ class PiecesBar(BoxLayout):
 
     def __init__(self, board: Board, **kwargs) -> None:
         super().__init__(**kwargs)
+        if not isinstance(board, Board):
+            raise ValueError(f':board: needs to be a Board object, not {type(board)}!')
+        self.board = board
         self.selected = None
         self.confirmed = None
-        self._widgets = []
         self.pieces_set = PiecesBar.generate_pieces_set()
-        self.board = board
+        self.widgets = []
         self.add_pieces()
-        self.confirm_button = Button(text='Confirm')
+        self.confirm_button = Button(text='Confirm', size_hint_x=None)
         self.confirm_button.bind(on_release=self.confirm)
         self.add_widget(self.confirm_button)
-        self.board.pieces_bar = self
-        self.board.disabled = False
-        self.board.initialise_buttons()
-        self.board.first_move()
         Logger.debug("Init: Initialised new PieceBar object.")
 
     def add_pieces(self):
         for piece in self.pieces_set:
             widget = Cell(piece)
             widget.bind(on_release=self.select)
-            self._widgets.append(widget)
+            self.widgets.append(widget)
             self.add_widget(widget)
 
     def confirm(self, touch) -> None:
@@ -221,18 +225,29 @@ class PiecesBar(BoxLayout):
         :param touch:   The confirm button
         :return:        None
         """
-        if self.selected is not None and self.confirmed is None:
-            self.confirmed = self.selected
-            self.selected.color = Color.CONFIRMED_TINT.value
-            self.selected = None
-            self.board.change_player()
-            if self.board.game_mode == GameMode.SINGLE_PLAYER:
-                self.board.computer_move()
+        if isinstance(touch, Piece):
+            assert self.confirmed is None and self.selected is None
+            for cell in self.widgets:
+                if cell.piece == touch:
+                    self.confirmed = cell
+                    cell.color = Color.CONFIRMED_TINT.value
+                    break
+            if self.confirmed is None:
+                raise ValueError('Invalid piece to confirm.')
+            self.pieces_set.remove(touch)
+        else:
+            if self.selected is not None and self.confirmed is None:
+                self.confirmed = self.selected
+                self.selected.color = Color.CONFIRMED_TINT.value
+                self.pieces_set.remove(self.selected.piece)
+                self.selected = None
+                self.board.change_player()
+                if self.board.game_mode == GameMode.single_player:
+                    self.board.computer_move()
 
     def remove_confirmed(self):
         assert self.confirmed is not None
-        self._widgets.remove(self.confirmed)
-        self.pieces_set.remove(self.confirmed.piece)
+        self.widgets.remove(self.confirmed)
         self.remove_widget(self.confirmed)
         self.confirmed = None
 
@@ -248,18 +263,21 @@ class PiecesBar(BoxLayout):
             self.selected = touch
             touch.color = Color.SELECTED_TINT.value
 
+    def clear_all(self):
+        while len(self.widgets) > 0:
+            self.remove_widget(self.widgets[0])
+            del self.widgets[0]
+
     def reset(self) -> None:
         self.remove_widget(self.confirm_button)
+        self.clear_all()
         self.pieces_set = PiecesBar.generate_pieces_set()
-        while len(self._widgets) > 0:
-            self.remove_widget(self._widgets[0])
-            del self._widgets[0]
         self.add_pieces()
         self.add_widget(self.confirm_button)
         self.disabled = False
 
     def __len__(self):
-        return len(self._widgets)
+        return len(self.widgets)
 
 
 class EndMessage(Popup):