+++ src/piece.py
@@ -6,29 +6,44 @@ Contains the declarations of the Piece and Cell classes.
 """
 
 __all__ = ['Piece', 'Cell']
-__version__ = '0.0'
+__version__ = '0.3'
 __author__ = 'Eric G.D'
 
 import os
-from typing import Any, Union
+from typing import Any, Tuple, Union
 
 from kivy.uix.behaviors import ButtonBehavior
-from kivy.uix.image import Image
+from kivy.uix.image import AsyncImage
 
 
 class Piece:
-    BLANK_IMAGE = os.path.join('assets', 'blank.png')
+    """
+    Class Piece
+    -----------
+
+    Contains the data of every piece, including it's id number, attributes tuple and image representation
+    """
     NUM_OF_BITS = 4
     MAX_NUM = (2 ** NUM_OF_BITS) - 1
 
     def __init__(self, num: int) -> None:
+        self.__id = num
+        self.__attributes = Piece.get_attributes(num)
+        self.__image = os.path.join('assets', f'piece_{str(num).zfill(2)}.png')
+
+    @staticmethod
+    def get_attributes(num: int) -> Tuple[bool, ...]:
+        """
+        :param num:     The piece's ID
+        :return:        The attributes of the piece with the id :num: (Based on the :num:'s binary representation)
+        """
         if not isinstance(num, int):
-            raise TypeError(':num: needs to be an integer!')
+            raise TypeError(f':num: needs to be an int, not {type(num)}!')
         if not 0 <= num <= Piece.MAX_NUM:
             raise ValueError(f':num: needs to be between 0 and {Piece.MAX_NUM}')
-        self.__id = num
-        self.__attributes = tuple([bool(int(bit)) for bit in ('{0:0=%db}' % Piece.NUM_OF_BITS).format(num)])
-        self.__image = os.path.join('assets', f'piece_{str(num).zfill(2)}.png')
+        binary_str = format(num, 'b').zfill(Piece.NUM_OF_BITS)
+        attributes = [bit == '1' for bit in binary_str]
+        return tuple(attributes)
 
     def __eq__(self, other: Any) -> bool:
         return isinstance(other, Piece) and other.__id == self.__id
@@ -37,19 +52,23 @@ class Piece:
         return hash(self.__id)
 
     @property
-    def isLarge(self) -> bool:
+    def attributes(self) -> Tuple[bool, ...]:
+        return self.__attributes
+
+    @property
+    def is_large(self) -> bool:
         return self.__attributes[0]
 
     @property
-    def isRound(self) -> bool:
+    def is_round(self) -> bool:
         return self.__attributes[1]
 
     @property
-    def isHollow(self) -> bool:
+    def is_hollow(self) -> bool:
         return self.__attributes[2]
 
     @property
-    def isWhite(self) -> bool:
+    def is_white(self) -> bool:
         return self.__attributes[3]
 
     @property
@@ -61,18 +80,25 @@ class Piece:
         return self.__id
 
 
-class Cell(ButtonBehavior, Image):
+class Cell(ButtonBehavior, AsyncImage):
+    """
+    class Cell(ButtonBehavior, AsyncImage)
+    --------------------------------------
 
-    def __init__(self, piece: Union[int, Piece] = None):
+    Represents a cell in the game board and a button in PiecesBar
+    This class is essentially a graphic wrapper for Piece
+    """
+    BLANK_IMAGE = os.path.join('assets', 'blank.png')
+
+    def __init__(self, piece: Union[None, int, Piece] = None):
         ButtonBehavior.__init__(self)
-        Image.__init__(self)
+        AsyncImage.__init__(self)
         if isinstance(piece, int):
             piece = Piece(piece)
+        if not isinstance(piece, Piece) and piece is not None:
+            raise TypeError(f'{type(piece)} is not a valid type for :piece:!')
         self.__piece = piece
-        try:
-            self.source = piece.image if piece is not None else Piece.BLANK_IMAGE
-        except AttributeError:
-            raise TypeError(f'{type(piece)} is an invalid type for :piece:!')
+        self.source = piece.image if piece is not None else Cell.BLANK_IMAGE
 
     @property
     def piece(self) -> Piece:
@@ -82,7 +108,7 @@ class Cell(ButtonBehavior, Image):
     def piece(self, p: Union[None, int, Piece]) -> None:
         if p is None:
             self.__piece = p
-            self.source = Piece.BLANK_IMAGE
+            self.source = Cell.BLANK_IMAGE
         elif isinstance(p, int):
             if not 0 <= p <= Piece.MAX_NUM:
                 raise ValueError(f':p: needs to be between 0 and {Piece.MAX_NUM}')
@@ -92,4 +118,4 @@ class Cell(ButtonBehavior, Image):
             self.__piece = p
             self.source = self.__piece.image
         else:
-            raise TypeError(f'{type(p)} is an invalid type for :p:!')
+            raise TypeError(f'{type(p)} is not a valid type for :p:!')